+++ packages/medusa/src/api/routes/admin/auth/get-session.ts
@@ -24,10 +24,11 @@ import _ from "lodash"
  *     label: cURL
  *     source: |
  *       curl '{backend_url}/admin/auth' \
- *       -H 'Authorization: Bearer {api_token}'
+ *       -H 'x-medusa-access-token: {api_token}'
  * security:
  *   - api_token: []
  *   - cookie_auth: []
+ *   - jwt_token: []
  * tags:
  *   - Auth
  * responses: