+++ packages/medusa/src/api/routes/admin/order-edits/add-line-item.ts
@@ -42,12 +42,13 @@ import {
  *     label: cURL
  *     source: |
  *       curl -X POST '{backend_url}/admin/order-edits/{id}/items' \
- *       -H 'Authorization: Bearer {api_token}' \
+ *       -H 'x-medusa-access-token: {api_token}' \
  *       -H 'Content-Type: application/json' \
  *       --data-raw '{ "variant_id": "variant_01G1G5V2MRX2V3PVSR2WXYPFB6", "quantity": 3 }'
  * security:
  *   - api_token: []
  *   - cookie_auth: []
+ *   - jwt_token: []
  * tags:
  *   - Order Edits
  * responses: