+++ packages/pricing/src/loaders/container.ts
@@ -3,8 +3,8 @@ import * as defaultServices from "@services"
 
 import { LoaderOptions } from "@medusajs/modules-sdk"
 import { ModulesSdkTypes } from "@medusajs/types"
-import { asClass } from "awilix"
 import { loadCustomRepositories } from "@medusajs/utils"
+import { asClass } from "awilix"
 
 export default async ({
   container,
@@ -22,7 +22,12 @@ export default async ({
     moneyAmountService: asClass(defaultServices.MoneyAmountService).singleton(),
     priceSetService: asClass(defaultServices.PriceSetService).singleton(),
     ruleTypeService: asClass(defaultServices.RuleTypeService).singleton(),
+    priceSetMoneyAmountRulesService: asClass(
+      defaultServices.PriceSetMoneyAmountRulesService
+    ).singleton(),
     priceRuleService: asClass(defaultServices.PriceRuleService).singleton(),
+    priceSetRuleTypeService: asClass(defaultServices.PriceSetRuleTypeService).singleton(),
+    priceSetMoneyAmountService : asClass(defaultServices.PriceSetMoneyAmountService).singleton(),
   })
 
   if (customRepositories) {
@@ -51,8 +56,17 @@ function loadDefaultRepositories({ container }) {
     ruleTypeRepository: asClass(
       defaultRepositories.RuleTypeRepository
     ).singleton(),
+    priceSetMoneyAmountRulesRepository: asClass(
+      defaultRepositories.PriceSetMoneyAmountRulesRepository
+    ).singleton(),
     priceRuleRepository: asClass(
       defaultRepositories.PriceRuleRepository
+    ).singleton(),
+    priceSetRuleTypeRepository: asClass(
+      defaultRepositories.PriceSetRuleTypeRepository
+    ).singleton(),
+    priceSetMoneyAmountRepository: asClass(
+      defaultRepositories.PriceSetMoneyAmountRepository
     ).singleton(),
   })
 }