+++ www/apps/api-reference/specs/admin/code_samples/Shell/admin_store_payment-providers/get.sh
@@ -1,2 +1,2 @@
 curl '{backend_url}/admin/store/payment-providers' \
--H 'Authorization: Bearer {api_token}'
+-H 'x-medusa-access-token: {api_token}'